+++ changelog.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 2023-01-16
+
+### Changed
+- add new field to serializer
+
+### Added
+- can search machine by name and location
+
+
 ## 2023-01-15
 
 ### Changed